% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastdm.R
\name{fastRdm}
\alias{fastRdm}
\title{Fit diffusion model}
\usage{
fastRdm()
}
\arguments{
\item{formula}{model formula (see details).}

\item{dat}{data.frame or named list consisting of many data.frames. Each data.frame must contain the following variables:
a numeric vector TIME, containing RT from a respective trial
a boolean/integer vector RESPONSE, containing the given response FALSE/0 or TRUE/1 from the respective trial}

\item{method}{character, optimization method: Kolmogorov-Smirnov (ks), Maximum Likelihood (ml), or Chi-Square (cs).}

\item{precision}{numeric, approximate precision of parameter estimates in decimals.}

\item{wd_fastdm}{character, working directory containing fast-dm executable (defaults to 'path').}

\item{res_prefix}{character, prefix of temporary data files.}

\item{removeTempFiles}{logical, should the temporary files used by fast-dm should be kept (FALSE) or removed (TRUE).}
}
\value{
data.frame, parameter estimates and fit statistics for each fitted RT distribution.
}
\description{
Specify diffusion model and fit it to response time (RT) data using the fast-dm library (v30.2)
}
\examples{
\dontrun{
path <- "./inst/binaries_30_2"
dat <- rdm(100, a=1, zr=.5, v=1.25, t0=0, N = 10)
out <- fastdm(~ zr*.5 + szr*0 + sv*0 + st0*0 + d*0, dat, method = "ks", precision = 3, res_prefix = "ID")
}
}
